requirements.txt
Distutils2==1.0a4
Django==1.7
anyjson==0.3.3
bunch==1.0.1
bzr==2.2.3
configobj==4.7.2
lxml==3.0.1
mock==1.0.1
nose==1.3.0
python-dateutil==2.1
pytz==2012h
pyzmq==2.2.0.1
pyzmq-static==2.2
redis==2.7.6
requests==1.2.3
six==1.8.0
springpython==1.3.0.RC1
urllib3==1.5
zato-client==1.2.1
zato-common==1.2.1
